fix : 최근 본 게시물 수정

+++ components/MainCard/index.tsx
@@ -1,7 +1,9 @@
 import Image from "next/image";
 import Link from "next/link";
+import { useRouter } from "next/router";
 import { useEffect, useMemo, useState } from "react";
 
+import useCookie from "@/hooks/useCookie";
 import { deleteBookMarks, postBookMarks } from "@/lib/api/bookmarks";
 
 function MainCard({ content, is }: { content: any; is: string }) {
@@ -14,6 +16,8 @@ function MainCard({ content, is }: { content: any; is: string }) {
       return "mobile:min-w-264 border border-line-02 w-11/12 h-[310px] block bg-white rounded-16 m-auto overflow-hidden";
     }
   }, [is]);
+  const router = useRouter();
+  const accessToken = useCookie("accessToken");
 
   const gender = useMemo(() => {
     switch (content.gender) {
@@ -29,14 +33,19 @@ function MainCard({ content, is }: { content: any; is: string }) {
   }, [content.gender]);
 
   const handleFavor = async () => {
-    if (favor) {
-      await postBookMarks(content.id);
-      setFavorCount((prev: number) => prev + 1);
+    if (accessToken !== null) {
+      if (favor) {
+        await postBookMarks(content.id);
+        setFavorCount((prev: number) => prev + 1);
+      } else {
+        await deleteBookMarks(content.id);
+        setFavorCount((prev: number) => prev - 1);
+      }
+      setFavor(prev => !prev);
     } else {
-      await deleteBookMarks(content.id);
-      setFavorCount((prev: number) => prev - 1);
+      alert("로그인 후 이용해 주시기 바랍니다.");
+      router.push("/login");
     }
-    setFavor(prev => !prev);
   };
 
   useEffect(() => {
@@ -105,10 +114,10 @@ function MainCard({ content, is }: { content: any; is: string }) {
             )}
           </div>
 
-          <div className="text-text-white mt-16 text-16 leading-tight tablet:text-14 ">
+          <div className="mt-16 text-16 leading-tight text-text-white tablet:text-14 ">
             {content.title}
           </div>
-          <div className="text-text-white mt-1 text-14">{content.nickname}</div>
+          <div className="mt-1 text-14 text-text-white">{content.nickname}</div>
         </div>
       </div>
